pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legacylens"
version = "0.1.0"
description = "RAG system for querying legacy COBOL codebases via natural language"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "pinecone>=6.0",
    "openai>=1.0",
    "tiktoken>=0.7",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "jinja2>=3.1",
    "typer>=0.14",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pytest>=8.0",
    "anthropic>=0.40",
]

[project.scripts]
legacylens = "legacylens.cli:app"

[tool.setuptools.packages.find]
include = ["legacylens*"]
